+++ src/tools/manifestConverter.js
@@ -1,5 +1,4 @@
 import {grabPath} from "./compatibility.js";
-import template from "../files/manifestTemplate.json" with {type: "json"};
 
 import path from "path";
 import os from "os";
@@ -7,11 +6,8 @@ import os from "os";
 const osCurrent = os.platform();
 
 async function addOSSpecArguments(original) {
-    console.log('run')
     if (osCurrent === "win32") {
-        original.arguments.jvm.push(
-            "-XX:HeapDumpPath=MojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump",
-        );
+        original.arguments.jvm.push("-XX:HeapDumpPath=MojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump",);
         original.arguments.jvm.push("-Dos.version=10.0");
     } else if (osCurrent === "darwin") {
         // Unsure if needed on newer Mac versions. require hardware to test
@@ -30,12 +26,7 @@ async function convertAssets(original) {
             let obj = {
                 origin: `https://resources.download.minecraft.net/${objs[idx].hash.substring(0, 2)}/${objs[idx].hash}`,
                 sha1: objs[idx].hash,
-                destination: path.join(
-                    CAULDRON_PATH,
-                    "assets",
-                    "objects",
-                    objs[idx].hash.substring(0, 2),
-                ),
+                destination: path.join(CAULDRON_PATH, "assets", "objects", objs[idx].hash.substring(0, 2),),
                 fileName: objs[idx].hash,
             };
             newData.push(obj);
@@ -55,13 +46,7 @@ async function convertLegacyAssets(original) {
             let obj = {
                 origin: `https://resources.download.minecraft.net/${objs[idx].hash.substring(0, 2)}/${objs[idx].hash}`,
                 sha1: objs[idx].hash,
-                destination: path.join(
-                    CAULDRON_PATH,
-                    "assets",
-                    "virtual",
-                    "legacy",
-                    splitPath.join("/"),
-                ),
+                destination: path.join(CAULDRON_PATH, "assets", "virtual", "legacy", splitPath.join("/"),),
                 fileName: fileName,
             };
             newData.push(obj);
@@ -91,8 +76,5 @@ async function convertPre16Assets(original) {
 }
 
 export {
-    addOSSpecArguments,
-    convertAssets,
-    convertLegacyAssets,
-    convertPre16Assets,
+    addOSSpecArguments, convertAssets, convertLegacyAssets, convertPre16Assets,
 };